- Modified classes in CArray.h (CGObList, CGTypedArray, CGPtrTypeArray, CGObArray, CGObSortArray) and derivatives (CCharRefArray, CRandGroupDef, CResourceQtyArray, CResourceRefArray, CResourceHash) to use size_t for array lengths and indices. Since '< 0' is no longer a valid method of checking if a container index is invalid, a new BadIndex() method returns a value (numeric_limits<size_t>::max()) that can be considered an 'invalid' index. For situations where we're only interested in checking for an item's existence, several Contains* methods have been added as shortcuts to "Find*() != BadIndex()". Also, several loops which depended on using "x >= 0" as their condition needed to be reworked. This comes with quite a number of changes across many code files to keep the data types consistent and to avoid undesirable side-effects, but I'm fairly confident that there shouldn't be any major issues.
- Added a couple of ASSERTs to help catch unexpected behaviour.
- Renamed MAX_SKILL to SKILL_MAX, and most of its usage with a direct call to g_Cfg.m_iMaxSkill. The idea is that SKILL_MAX should be used when dealing with SKILL_TYPE variables (there's an automatic type cast) and m_iMaxSkill should be used everywhere else. Also changed the type of CItemBase::m_iSkill to SKILL_TYPE since it's always having to be cast that way.
- Fixed possible memory access violation in CWorldThread::AllocUID due to a check being made in the order order, leading to the value of m_FreeOffset being validated only after it has already been used to access the array.

+++ 0.56.b/common/CArray.cpp
@@ -93,7 +93,7 @@ void CGObList::InsertAfter( CGObListRec * pNewRec, CGObListRec * pPrev )
 	m_iCount ++;
 }
 
-CGObListRec * CGObList::GetAt( int index ) const
+CGObListRec * CGObList::GetAt( size_t index ) const
 {
 	ADDTOCALLSTACK("CGObList::GetAt");
 	CGObListRec * pRec = GetHead();
@@ -158,7 +158,7 @@ CGObListRec * CGObList::GetTail() const
 	return m_pTail;
 }
 
-int CGObList::GetCount() const
+size_t CGObList::GetCount() const
 {
 	ADDTOCALLSTACK("CGObList::GetCount");
 	return m_iCount;
